+++ CHANGELOG.rst
@@ -15,7 +15,8 @@ Change Log
 [0.5.5] 2021-xx-yy
 -------------------
 - [ADDED] possibility to perform dc powerflow
-- [ADDED] a class to compute flows on whole time series without using grid2op.
+- [ADDED] a class to compute flows on whole time series when the Ybus does not change (see `TimeSerie`)
+- [ADDED] a class to compute flows on multiple contingencies, when Sbus does not change (see ).
 - [IMPROVED] running speed of Newton Raphson solvers with better filling of sparse matrices
 - [IMPROVED] upgrade to SuiteSparse `v5.10.1`
 - [IMPROVED] upgrade to eigen `3.4.0` (stable release)